+++ src/modules/CopyObjectString.cpp
@@ -1,7 +1,5 @@
 #include "modules/CopyObjectString.hpp"
 
-#ifndef GEODE_IS_IOS
-
 bool CopyObjectString::onToggled(bool state) {
     return true;
 }
@@ -30,5 +28,3 @@ void COSEditorUI::doPasteObjects(bool withColor) {
         Notification::create("Pasted Objects from Clipboard", NotificationIcon::Info)->show();
     }
 }
-
-#endif